requirements.txt
einops==0.6.0
h5py==3.7.0
ignite==1.1.0
imageio==2.19.3
matplotlib==3.6.2
numpy==1.21.5
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.7.3
timm==0.6.12
torch==1.7.0
torchvision==0.8.1
wandb==0.13.10
